installing git in server

+++ 07_git_installation.sh
@@ -10,22 +10,32 @@ then
     exit 1
 fi
 
-dnf install git -y
+dnf list installed git
+if [ $? -e 0]
+then
+    echo "git already installed and up-to-date"
+else
+    dnf install git -y
 
-if [ $? -ne 0 ]
-then   
-    echo "git installation failed"
-    exit 1
-else   
-    echo "git installation is success"
+    if [ $? -ne 0 ]
+    then   
+        echo "git installation failed"
+        exit 1
+    else   
+        echo "git installation is success"
 fi
 
-dnf install ansible -y
 
-if [ $? -ne 0 ]
-then   
-    echo "ansible installation failed"
-    exit 1
-else   
-    echo "ansible installation is success"
+dnf list installed git
+if [ $? -e 0]
+then
+    echo "ansible already installed and up-to-date"
+else
+    dnf install ansible -y
+    if [ $? -ne 0 ]
+    then   
+        echo "ansible installation failed"
+        exit 1
+    else   
+        echo "ansible installation is success"
 fi
